+++ lib/danger/ci_source/buildkite.rb
@@ -14,8 +14,15 @@ module Danger
   #
   # ### Token Setup
   #
+  # #### GitHub
+  #
   # As this is self-hosted, you will need to add the `DANGER_GITHUB_API_TOKEN` to your build user's ENV. The alternative
   # is to pass in the token as a prefix to the command `DANGER_GITHUB_API_TOKEN="123" bundle exec danger`.
+  #
+  # #### GitLab
+  #
+  # As this is self-hosted, you will need to add the `DANGER_GITLAB_API_TOKEN` to your build user's ENV. The alternative
+  # is to pass in the token as a prefix to the command `DANGER_GITLAB_API_TOKEN="123" bundle exec danger`.
   #
   class Buildkite < CI
     def self.validates_as_ci?(env)
@@ -36,7 +43,7 @@ module Danger
     end
 
     def supported_request_sources
-      @supported_request_sources ||= [Danger::RequestSources::GitHub]
+      @supported_request_sources ||= [Danger::RequestSources::GitHub, Danger::RequestSources::GitLab]
     end
   end
 end